client/requirements.txt
Flask
tabulate
pandas
mysql-connector
requests
